examples/python/v2/clients/requests/pyproject.toml
[project]
name = "x402-requests-client-example"
version = "0.1.0"
description = "Example of using requests client with x402 payment handling (V2)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "eth-account>=0.8.0",
    "python-dotenv>=1.0.0",
    "x402"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
package = false

[tool.uv.sources]
x402 = { path = "../../../../../python/x402", editable = true }
